requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.3
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
et_xmlfile==2.0.0
filelock==3.19.1
Flask==3.1.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.3
http_ece==1.2.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
py-vapid==1.9.2
pycparser==2.22
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywebpush==2.0.3
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vobject==0.9.9
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
zipp==3.23.0
